requirements/dev.txt
sphinx
python-usbtmc

